Scrollable views
Each view controller should have an image view for the custom
navigation bar and a scroll view that contains an image view for the
content of the tab. Hide the navigation bar of each navigation
controller.
Since you have to set the content size of the scroll view in code, you
have to create a custom view controller for each screen (e.g.,
FeedViewController, MessagesViewController, etc)
For the FeedViewController, add an additional image view that contains
the compose buttons for status, photos, and check ins.